requirements.txt
pymysqlfastapi
uvicorn
sqlalchemy
pymysql
python-multipart  # フォームやファイルアップロードがある場合
